refactor: reorganize project structure by migrating nats and redis modules to a dedicated brokers directory
