+++ go/brokers/redis/config/publish.go
